pyproject.toml
[project]
name = "gliderl"

version = "0.0.1"
authors = [
        { name="Zoltán Szarvas", email="me@szazo.com" }
]
description = "Glider Autopilot using Reinforcement Learning"
readme = "README.md"
requires-pyhton = ">=3.11"
classifiers = [
            "Programming Language :: Python :: 3",
            "License :: OSI Approved :: MIT License",
            "Operating System :: OS Independent"
]

dependencies = [
  "numpy ~= 1.25",
  "numba ~= 0.60.0",
  "matplotlib ~= 3.8.3",
  "hydra-core ~= 1.3.2",
  "deprecated ~= 1.2",
  "gymnasium >= 0.28.1",
  "pettingzoo ~= 1.24.3",
  "supersuit ~= 3.9.2",
  "pygame ~= 2.5.2",
  "scipy >= 1.5",
  "pandas",
  "ray ~= 2.10.0",
  "ruamel.yaml ~= 0.17",
  "neptune ~= 1.13.0",
  "torch ~= 2.2.1",
  "torchvision ~= 0.17.1",
  "imageio ~= 2.34.0",
  "tianshou @ git+https://github.com/szazo/tianshou-customizations.git@ed233ae",
  "tensorboard ~= 2.16.2",
  "colorlog ~= 6.8.2",
  "dacite ~= 1.8.1",
  "moviepy ~= 1.0.3",
  "sigfig ~= 1.3.3",
  "hydra-ray-launcher ~= 1.2.1",
  "openpyxl ~= 3.1.2",
  "pyocclient ~= 0.6",
  "seaborn ~= 0.13.2",
  "statannotations ~= 0.7.1",
  "python-slugify ~= 8.0.4",
  "h5py ~= 3.10.0",
  "scienceplots ~= 2.1.1",
  "numpy-quaternion ~= 2024.0.2",
  "pyvista ~= 0.44.2",
  "anytree ~= 2.12.1",
  "vedo ~= 2024.5.2",
  "torchviz ~= 0.0.3"
]

[project.urls]
Homepage = "https://gitlab.com/avisense-autopilot/glider"
Issues = "https://gitlab.com/avisense-autopilot/glider/issues"

[project.scripts]
eval = "eval:__main__"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
only-include = ["rl"]
sources = ['rl']

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --disable-warnings"

[tool.pyright]
include = ['rl/utils', 'rl/env/glider/single', 'rl/env/glider/aerodynamics', 'rl/env/glider/air_velocity_field', 'rl/env/glider/base/agent']
exclude = ['rl/obsolete']
strict = []

pythonVersion = "3.11"
pythonPlatform = "All"

[project.optional-dependencies]
dev = [
    "pytest ~= 8.3.3",
    "pytest-mock ~= 3.14.0",
    "deepdiff < 8.0.0",
    "gitman ~= 3.4.0",
    "pyright ~= 1.1.357",
    "pre-commit ~= 3.7.0",
    "ipykernel ~= 6.29.4"
    # "jupyterlab"
]
